web/explain: pass all of the matcher's matchgroups to the explain template
The idea is to send each group to the template, and let the user navigate
between groups with a prev/next button. Each time, a certain group will be
displayed, with the 'shell' group being the first. So when the page loads,
the user is going to see explanations of shell syntax, and clicking the next
button will move to the first group of command explanations.

Suppose we have the command 'a | b', then first we're going to see what '|'
means, then clicking next will show what 'a' means, and another click will show
what 'b' means.

We might want to show everything in case there are <5 (or something like that)
lines.