web/explain: connect matches with the same help text to a single box
Previously we displayed multiple help boxes for (non adjacent) boxes, because
only adjacent matches were merged. This changes that so all matches that share
the same help text will connect to the same box, and thus help boxes with the
same text will not be duplicated.

For now, lines simply cross one another, but we might want to add a visual to
make it clearer when they meet, such as dotting, or a half circle.