web: change url format to /explain?cmd=... (fixes #7)
Now that we support complex commands, it no longer makes sense to have urls of
the form '/explain/<prog>..' since a command might involve multiple programs.

Instead, we change the url format the more simplistic form of /explain?cmd=,
which should also make it easier to consume and doesn't require manipulating
the query, like we used to do with client code.

The old url will remain for now.