web/explain: rework the connection between the command and help
The previous behaviour associated different parts of a group with the same help
id to prevent duplication of help boxes. But then it was possible for a single
command to span two groups, e.g. 'a -b && a -c' which would result in two
groups, but the second group would consist of only '-c'. This would mess up the
navigation because the second group would not show anything for the second 'a'.

Now each group is given a different name, so the first 'a' would be command0,
and the second would be command1, but we keep a reference for each span to its
help <pre> and handle it accordingly.