es.js: fix an issue when moving between groups
When moving groups we hide/show some help boxes which may cause the page height
to shrink. This caused an issue when redrawing the lines and a misclaculating
the top of the canvas.

This is triggerable for example with the cut.. example, scrolling down a bit
and then moving to the next group.