matcher: ensure that the program part of a command isn't marked as unknown
If a man page has no synopsis, it could be mistaken to be an unknown part
of the command, even though we have a matching man page. So workaround that
by setting a default 'no synopsis found' so everything else works as expected.

Maybe we're better off with not including these man pages in the db at all?
They might still contain options, but the one I've seen (puppet) doesn't and
I don't feel like messing with that right now.